Added Imgur support, fixed 2+ posting bug
Two Image handler classes handle downloading and uploading from/to
imgur. PostContainer holds both a ParseObject and the associated bitmap
with this object. PostAdapter now uses objectID strings to keep track
of what posts it should display, and gets the info for each post from
the PostContainer map in ParseHandler.